requirements.txt
altair
numpy
pandas
pydeck
streamlit
seaborn
matplotlib.pyplot
LinearRegression